requirements.txt
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==3.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
numpy==1.24.2
pandas==1.5.3
pyTelegramBotAPI==4.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
requests==2.28.2
six==1.16.0
urllib3==1.26.14
yarl==1.8.2
